+++ ru.taximaxim.codekeeper.core/src/main/java/ru/taximaxim/codekeeper/core/parsers/antlr/statements/ms/CreateMsFunction.java
@@ -21,7 +21,6 @@ import java.util.List;
 import org.antlr.v4.runtime.CommonTokenStream;
 import org.antlr.v4.runtime.ParserRuleContext;
 
-import ru.taximaxim.codekeeper.core.DatabaseType;
 import ru.taximaxim.codekeeper.core.model.difftree.DbObjType;
 import ru.taximaxim.codekeeper.core.parsers.antlr.expr.launcher.MsFuncProcTrigAnalysisLauncher;
 import ru.taximaxim.codekeeper.core.parsers.antlr.generated.TSQLParser.Assembly_specifierContext;
@@ -92,8 +91,7 @@ public class CreateMsFunction extends BatchContextProcessor {
             MsClrFunction func = new MsClrFunction(name, assembly,
                     assemblyClass, assemblyMethod);
 
-            addDepSafe(func, Arrays.asList(assemblyCtx.assembly_name),
-                    DbObjType.ASSEMBLY, DatabaseType.MS);
+            addDepSafe(func, Arrays.asList(assemblyCtx.assembly_name), DbObjType.ASSEMBLY);
 
             for (Function_optionContext option : ctx.function_option()) {
                 func.addOption(getFullCtxText(option));