+++ ru.taximaxim.codekeeper.core/src/main/java/ru/taximaxim/codekeeper/core/parsers/antlr/statements/pg/CreateFtsTemplate.java
@@ -19,7 +19,6 @@ import java.util.List;
 
 import org.antlr.v4.runtime.ParserRuleContext;
 
-import ru.taximaxim.codekeeper.core.DatabaseType;
 import ru.taximaxim.codekeeper.core.model.difftree.DbObjType;
 import ru.taximaxim.codekeeper.core.parsers.antlr.QNameParser;
 import ru.taximaxim.codekeeper.core.parsers.antlr.generated.SQLParser.Create_fts_template_statementContext;
@@ -47,13 +46,11 @@ public class CreateFtsTemplate extends PgParserAbstract {
 
         if (ctx.init_name != null) {
             template.setInitFunction(getFullCtxText(ctx.init_name));
-            addDepSafe(template, getIdentifiers(ctx.init_name), DbObjType.FUNCTION, DatabaseType.PG,
-                    "(internal)");
+            addDepSafe(template, getIdentifiers(ctx.init_name), DbObjType.FUNCTION, "(internal)");
         }
 
         template.setLexizeFunction(getFullCtxText(ctx.lexize_name));
-        addDepSafe(template, getIdentifiers(ctx.lexize_name), DbObjType.FUNCTION, DatabaseType.PG,
-                "(internal, internal, internal, internal)");
+        addDepSafe(template, getIdentifiers(ctx.lexize_name), DbObjType.FUNCTION, "(internal, internal, internal, internal)");
 
         addSafe(getSchemaSafe(ids), template, ids);
     }