+++ ru.taximaxim.codekeeper.core/src/main/java/ru/taximaxim/codekeeper/core/parsers/antlr/statements/pg/CreateFunction.java
@@ -24,7 +24,6 @@ import org.antlr.v4.runtime.Token;
 
 import ru.taximaxim.codekeeper.core.Consts;
 import ru.taximaxim.codekeeper.core.Consts.FUNC_SIGN;
-import ru.taximaxim.codekeeper.core.DatabaseType;
 import ru.taximaxim.codekeeper.core.PgDiffUtils;
 import ru.taximaxim.codekeeper.core.model.difftree.DbObjType;
 import ru.taximaxim.codekeeper.core.parsers.antlr.AntlrParser;
@@ -170,7 +169,7 @@ public final class CreateFunction extends PgParserAbstract {
             } else if (action.SUPPORT() != null) {
                 List<ParserRuleContext> suppFuncIds = getIdentifiers(action.schema_qualified_name());
                 function.setSupportFunc(getFullCtxText(suppFuncIds));
-                addDepSafe(function, suppFuncIds, DbObjType.FUNCTION, DatabaseType.PG, FUNC_SIGN.INTERNAL.getName());
+                addDepSafe(function, suppFuncIds, DbObjType.FUNCTION, FUNC_SIGN.INTERNAL.getName());
             }
         }