+++ ru.taximaxim.codekeeper.core/src/main/java/ru/taximaxim/codekeeper/core/parsers/antlr/statements/pg/CreateUserMapping.java
@@ -17,7 +17,6 @@ package ru.taximaxim.codekeeper.core.parsers.antlr.statements.pg;
 
 import java.util.Arrays;
 
-import ru.taximaxim.codekeeper.core.DatabaseType;
 import ru.taximaxim.codekeeper.core.model.difftree.DbObjType;
 import ru.taximaxim.codekeeper.core.parsers.antlr.generated.SQLParser.Create_user_mapping_statementContext;
 import ru.taximaxim.codekeeper.core.parsers.antlr.generated.SQLParser.Define_foreign_optionsContext;
@@ -47,7 +46,7 @@ public class CreateUserMapping extends PgParserAbstract {
         String server = userMapping.identifier().getText();
 
         PgUserMapping usm = new PgUserMapping(userName.getText(), server);
-        addDepSafe(usm, Arrays.asList(userMapping.identifier()), DbObjType.SERVER, DatabaseType.PG);
+        addDepSafe(usm, Arrays.asList(userMapping.identifier()), DbObjType.SERVER);
 
         Define_foreign_optionsContext options = ctx.define_foreign_options();
         if (options != null) {