+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fixed an error when opening non-pgCodeKeeper project files in SQL editor.
 - Fixed a bug with defining a dependency in the body of a function in PG.
+- Added missing references to COLLATION objects in PostgreSQL.
 
 ## [9.0.0] - 2023-12-13